shared/requirements.txt
# 공통 라이브러리 의존성

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 데이터베이스
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# 인증
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# 검증
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 유틸리티
python-dotenv>=1.0.0
